+++ packages/terra-alert/src/Alert.jsx
@@ -36,9 +36,12 @@ const propTypes = {
    */
   customIcon: PropTypes.element,
   /**
-   * The status bar color to be used for an alert of type custom.
+   * Sets an author-defined class, to control the status bar color to be used for an alert of type custom.
+   *
+   * ![IMPORTANT](https://badgen.net/badge//IMPORTANT/CSS?icon=github)
+   * Adding `var(--my-app...` CSS variables is required for proper re-themeability when creating custom color styles _(see included examples)_.
    */
-  customStatusColor: PropTypes.string,
+  customColorClass: PropTypes.string,
   /**
    * Callback function triggered when Dismiss button is clicked. The presence of this prop will cause the Dismiss button to be included on the alert.
    */
@@ -66,7 +69,7 @@ const defaultProps = {
   action: null,
   children: '',
   customIcon: null,
-  customStatusColor: '',
+  customColorClass: 'custom-default-color',
   onDismiss: null,
   title: '',
   type: AlertTypes.ALERT,
@@ -107,7 +110,7 @@ const Alert = ({
   children,
   title,
   customIcon,
-  customStatusColor,
+  customColorClass,
   onDismiss,
   action,
   ...customProps
@@ -121,30 +124,23 @@ const Alert = ({
     type,
     'narrow',
     attributes.className,
+    { [`${customColorClass}`]: type === AlertTypes.CUSTOM },
   ]);
   const wideAlertClassNames = cx([
     'alert-base',
     type,
     'wide',
     attributes.className,
+    { [`${customColorClass}`]: type === AlertTypes.CUSTOM },
   ]);
 
   let actionsSection = '';
   let dismissButton = '';
-  const outerDivStyle = {};
   let alertSectionClassName = cx('section');
   let actionsClassName = cx('actions');
   let bodyClassNameForNarrowParent = cx(['body', 'body-std']);
 
   if (type === AlertTypes.CUSTOM) {
-    // For custom alert, there is no color assigned to the box-shadow style since it is to be specified
-    // in the customStatusColor prop.  The box-shadow style is defined in CSS in order to get the
-    // bidirectionality via the mixin.  As per the W3C spec if the box-shadow does not have the color
-    // defined, it will use the prevailing color style, so setting that here. But then we need to set
-    // the color style for the alert content so that it doesn't pick up the custom status color. We
-    // will allow the icon to pick up the color style so that Terra icons will match the color of the
-    // status bar.
-    outerDivStyle.color = customStatusColor;
     alertSectionClassName = cx(['section', 'section-custom']);
     actionsClassName = cx(['actions', 'actions-custom']);
   }
@@ -171,30 +167,28 @@ const Alert = ({
 
   return (
     <ResponsiveElement
-      responsiveTo="parent"
-      defaultElement={(
-        <div {...attributes} className={narrowAlertClassNames} style={outerDivStyle}>
+      tiny={(
+        <div {...attributes} className={narrowAlertClassNames}>
           <div className={bodyClassNameForNarrowParent}>
             {getAlertIcon(type, customIcon)}
             {alertMessageContent}
           </div>
           {actionsSection}
         </div>
-)}
-      tiny={(
-        <div {...attributes} className={wideAlertClassNames} style={outerDivStyle}>
+      )}
+      small={(
+        <div {...attributes} className={wideAlertClassNames}>
           <div className={cx(['body', 'body-std'])}>
             {getAlertIcon(type, customIcon)}
             {alertMessageContent}
           </div>
           {actionsSection}
         </div>
-)}
+      )}
     />
 
   );
 };
-
 Alert.propTypes = propTypes;
 Alert.defaultProps = defaultProps;
 Alert.contextTypes = contextTypes;