+++ packages/terra-alert/src/terra-dev-site/test/alert/SmallContainer.test.jsx
@@ -1,7 +1,11 @@
 import React from 'react';
 import Button from 'terra-button';
+import classNames from 'classnames/bind';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import Alert from 'terra-alert/lib/Alert';
+import styles from './AlertTestCommon.scss';
+
+const cx = classNames.bind(styles);
 
 class SmallContainerExample extends React.Component {
   constructor(props) {
@@ -26,7 +30,7 @@ class SmallContainerExample extends React.Component {
     const alert2EndTagStr = '</Alert>';
     const alert2Text = 'This is an information alert';
     return (
-      <div style={{ width: '500px', border: '1px solid black', padding: '5px' }}>
+      <div className={cx('content-wrapper')}>
         <div dir="ltr">
           <h3>
 Basic Alert of type warning with action button in a small container (500px wide) that will cause the action button section of the alert