+++ packages/terra-divider/src/terra-dev-site/test/divider/DefaultDivider.test.jsx
@@ -1,11 +1,15 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 
 import Divider from '../../../Divider';
+import styles from './DividerTestCommon.module.scss';
+
+const cx = classNames.bind(styles);
 
 export default () => (
   <div>
-    <div style={{ height: '2rem', background: '#40E0D0' }} />
+    <div className={cx('separator')} />
     <Divider id="divider" />
-    <div style={{ height: '2rem', background: '#40E0D0' }} />
+    <div className={cx('separator')} />
   </div>
 );