+++ packages/terra-form-radio/src/terra-dev-site/doc/example/field/ControlledRadioField.jsx
@@ -4,6 +4,10 @@ import React from 'react';
 import Radio from 'terra-form-radio/lib/Radio';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import RadioField from 'terra-form-radio/lib/RadioField';
+import classNames from 'classnames/bind';
+import styles from './RadioFieldCommon.module.scss';
+
+const cx = classNames.bind(styles);
 
 export default class extends React.Component {
   constructor(props) {
@@ -32,7 +36,7 @@ export default class extends React.Component {
     let errorMessage = this.state.toggleInvalid ? 'All options are now invalid' : undefined;
     let isInvalid = false;
 
-    if (this.state.selectedAnswer === 'mcdonalds') {
+    if (this.state.selectedAnswer === 'monday') {
       errorMessage = 'Invalid option selected.';
       isInvalid = true;
     } else if (this.state.selectedAnswer === undefined) {
@@ -43,19 +47,23 @@ export default class extends React.Component {
       <div>
         <div>
           <RadioField
-            legend="Which Type of Meal are you looking for?"
+            legend="What is your favorite day of the week?"
             help="This cannot be changed when submitted"
             isInvalid={this.state.toggleInvalid || isInvalid}
             error={errorMessage}
             required
           >
-            <Radio id="chicken-meal" name="meal" labelText="Chicken" onChange={this.handleOnChange} value="chicken" />
-            <Radio id="salmon-meal" name="meal" labelText="Salmon" onChange={this.handleOnChange} value="salmon" />
-            <Radio id="mcdonalds-meal" name="meal" labelText="McDonalds (Not a valid choice)" onChange={this.handleOnChange} value="mcdonalds" />
+            <Radio id="sunday" name="weekday" labelText="Sunday" onChange={this.handleOnChange} value="sunday" />
+            <Radio id="monday" name="weekday" labelText="Monday (Not a valid choice)" onChange={this.handleOnChange} value="monday" />
+            <Radio id="tuesday" name="weekday" labelText="Tuesday" onChange={this.handleOnChange} value="tuesday" />
+            <Radio id="wednesday" name="weekday" labelText="Wednesday" onChange={this.handleOnChange} value="wednesday" />
+            <Radio id="thursday" name="weekday" labelText="Thursday" onChange={this.handleOnChange} value="thursday" />
+            <Radio id="friday" name="weekday" labelText="Friday" onChange={this.handleOnChange} value="friday" />
+            <Radio id="saturday" name="weekday" labelText="Saturday" onChange={this.handleOnChange} value="saturday" />
           </RadioField>
         </div>
         <hr />
-        <button style={{ marginBottom: '5px' }} type="button" aria-label="Toggle Invalid Status" onClick={this.handleOnClick}>Toggle Invalid Status</button>
+        <button className={cx('radio-button-wrapper')} type="button" aria-label="Toggle Invalid Status" onClick={this.handleOnClick}>Toggle Invalid Status</button>
       </div>
     );
   }