+++ packages/terra-form-radio/src/terra-dev-site/doc/example/field/InlineRadioField.jsx
@@ -4,6 +4,11 @@ import React from 'react';
 import Radio from 'terra-form-radio/lib/Radio';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import RadioField from 'terra-form-radio/lib/RadioField';
+import classNames from 'classnames/bind';
+import styles from './RadioFieldCommon.module.scss';
+
+const cx = classNames.bind(styles);
+
 
 export default class extends React.Component {
   constructor(props) {
@@ -50,8 +55,8 @@ export default class extends React.Component {
 
     return (
       <div>
-        <button style={{ marginBottom: '5px' }} id="inline" type="button" aria-label="Toggle Inline" onClick={this.handleOnClick}>Toggle Inline</button>
-        <button style={{ marginBottom: '5px' }} id="invalid" type="button" aria-label="Toggle Inline" onClick={this.handleOnClick}>Toggle Invalid Status</button>
+        <button className={cx('radio-button-wrapper')} id="inline" type="button" aria-label="Toggle Inline" onClick={this.handleOnClick}>Toggle Inline</button>
+        <button className={cx('radio-button-wrapper')} id="invalid" type="button" aria-label="Toggle Inline" onClick={this.handleOnClick}>Toggle Invalid Status</button>
         <hr />
         <div>
           <RadioField