+++ packages/terra-form-select/src/terra-dev-site/doc/example/SelectField.jsx
@@ -1,9 +1,13 @@
 import React from 'react';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import SelectField from 'terra-form-select/lib/SelectField';
+import classNames from 'classnames/bind';
+import styles from './FormSelectDocCommon.module.scss';
+
+const cx = classNames.bind(styles);
 
 const SelectFieldExample = () => (
-  <SelectField label="T-shirt size" placeholder="Select a size" selectId="tshirt-size-field" required style={{ maxWidth: '300px' }}>
+  <SelectField label="T-shirt size" placeholder="Select a size" selectId="tshirt-size-field" className={cx('form-select')}>
     <SelectField.Option value="xSmall" display="Extra Small" />
     <SelectField.Option value="small" display="Small" />
     <SelectField.Option value="medium" display="Medium" />