+++ packages/terra-button/CHANGELOG.md
@@ -3,6 +3,16 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Moved v2 Upgrade guide that existed in the wiki to the upgrade guide doc
+
+### Changed
+* Removed all inline styles from Dev-site components and implemented the same styles using external css.
+
+3.17.0 - (June 18, 2019)
+------------------
+### Added
+* Added missing WDIO screenshots
 
 3.13.1 - (June 12, 2019)
 ------------------