+++ packages/terra-form-fieldset/CHANGELOG.md
@@ -3,6 +3,13 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Removed all inline styles from Dev-site components and implemented the same styles using external css.
+
+2.15.0 - (June 18, 2019)
+------------------
+### Removed
+* Removed stale WDIO screenshots
 
 2.14.0 - (June 11, 2019)
 ------------------