+++ packages/terra-form-select/CHANGELOG.md
@@ -3,6 +3,18 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Removed all inline styles from Dev-site components and implemented the same styles using external css.
+
+5.17.0 - (June 18, 2019)
+------------------
+### Fixed
+* Additional null check added to private menu to prevent error from empty query
+
+### Changed
+* Added aria roles for optgroups
+* Added `required` prop
+* Added `isSelect` type on Select component to help with component identity
 
 5.16.0 - (June 11, 2019)
 ------------------