+++ packages/terra-paginator/CHANGELOG.md
@@ -3,6 +3,16 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Removed all inline styles from Dev-site components and implemented the same styles using external css.
+
+2.15.0 - (June 18, 2019)
+------------------
+### Changed
+* Updated to use the new responsive element breakpoints
+
+### Fixed
+* Fixed pageSet logic to support total pages less than max allowed
 
 2.14.1 - (June 12, 2019)
 ------------------