+++ packages/terra-spacer/CHANGELOG.md
@@ -3,6 +3,13 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Removed all inline styles from Dev-site components and implemented the same styles using external css.
+
+3.12.0 - (June 18, 2019)
+------------------
+### Changed
+* Minor dependency bump
 
 3.11.1 - (June 12, 2019)
 ------------------