+++ packages/terra-status/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Removed all inline styles from Dev-site components and implemented the same styles using external css.
+
+4.7.0 - (June 18, 2019)
+------------------
+### Changed
+* Minor dependency bump
 
 4.6.0 - (June 11, 2019)
 ------------------