+++ packages/terra-toggle-button/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Removed all inline styles from Dev-site components and implemented the same styles using external css.
+
+3.13.0 - (June 18, 2019)
+------------------
+### Changed
+* Minor dependency bump
 
 3.12.1 - (June 12, 2019)
 ------------------